+++ CRUD_18_03/Application/DTOs/ResultadoEvaluacion/ResultadoEvaluacionDto.cs
@@ -3,7 +3,7 @@ namespace CRUD_18_03.Application.DTOs.ResultadoEvaluacion;
 public class ResultadoEvaluacionDto
 {
     public Guid Id { get; set; }
-    public double ScoreTotal { get; set; }
+    public decimal ScoreTotal { get; set; }
     public string Recomendacion { get; set; } = string.Empty;
     public string ResumenIA { get; set; } = string.Empty;
     public string BrechasDetectadas { get; set; } = string.Empty;
